refactor: consolidate CFDI stamping methods in CommonController and update PaymentEntry and SalesInvoice overrides
